Have seperate inlining threshold for functions using new method
Seems to have a 5-10% compile time impact. Fixes some regressions in problem BaseBenchmarks, some due to overinlining (when there are no cycles) and some due underinlining (when there are cycles). Still a 5% regression in go and a 10% one in simplex. Also no improvements really, laplacian I think will have one if 80 is changed to 75.